Fixing clang-format problems
- flyby: rename apex_sample_value.hpp
