+++ libs/performance_counters/src/manage_counter_type.cpp
@@ -135,8 +135,8 @@ namespace hpx { namespace performance_counters {
     // shutdown.
     counter_status install_counter_type(std::string const& name,
         hpx::util::function_nonser<std::int64_t(bool)> const& counter_value,
-        std::string const& helptext, std::string const& uom,
-        counter_type type, error_code& ec)
+        std::string const& helptext, std::string const& uom, counter_type type,
+        error_code& ec)
     {
         using hpx::util::placeholders::_1;
         using hpx::util::placeholders::_2;
