+++ libs/performance_counters/src/performance_counter_set.cpp
@@ -11,8 +11,8 @@
 #include <hpx/functional/bind.hpp>
 #include <hpx/lcos/future.hpp>
 #include <hpx/performance_counters/counters.hpp>
-#include <hpx/performance_counters/performance_counter_set.hpp>
 #include <hpx/performance_counters/performance_counter.hpp>
+#include <hpx/performance_counters/performance_counter_set.hpp>
 #include <hpx/runtime/get_locality_id.hpp>
 #include <hpx/runtime/launch_policy.hpp>
 #include <hpx/runtime_fwd.hpp>
