+++ libs/performance_counters/src/server/arithmetics_counter_extended.cpp
@@ -7,8 +7,8 @@
 #include <hpx/config.hpp>
 #include <hpx/performance_counters/counter_creators.hpp>
 #include <hpx/performance_counters/counters.hpp>
-#include <hpx/performance_counters/server/arithmetics_counter_extended.hpp>
 #include <hpx/performance_counters/performance_counter.hpp>
+#include <hpx/performance_counters/server/arithmetics_counter_extended.hpp>
 #include <hpx/runtime/components/derived_component_factory.hpp>
 #include <hpx/runtime/launch_policy.hpp>
 #include <hpx/runtime/runtime_fwd.hpp>
