+++ libs/performance_counters/src/server/statistics_counter.cpp
@@ -9,8 +9,8 @@
 #include <hpx/functional/bind_front.hpp>
 #include <hpx/performance_counters/counter_creators.hpp>
 #include <hpx/performance_counters/counters.hpp>
-#include <hpx/performance_counters/server/statistics_counter.hpp>
 #include <hpx/performance_counters/performance_counter.hpp>
+#include <hpx/performance_counters/server/statistics_counter.hpp>
 #include <hpx/runtime/actions/continuation.hpp>
 #include <hpx/runtime/agas/interface.hpp>
 #include <hpx/runtime/components/derived_component_factory.hpp>
